app/requirements.txt
flask==2.3.3
requests==2.31.0
pydantic==2.5.2
python-dotenv==1.0.0
langchain==0.0.354
transformers==4.36.2